+++ src/ubin/__init__.py
@@ -33,7 +33,7 @@ from .errors import (
     UbinTLSVerificationError,
 )
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 
 
 def open(source, *, name=None):